+++ src/video_player.c
@@ -472,7 +472,8 @@ static void video_player_clear_decode(VideoPlayer *player) {
     player->present_fps_valid = FALSE;
 }
 
-VideoPlayer* video_player_new(gint work_factor, gboolean force_text, gboolean force_sixel, gboolean force_kitty, gboolean force_iterm2) {
+VideoPlayer* video_player_new(gint work_factor, gboolean force_text, gboolean force_sixel, gboolean force_kitty,
+                              gboolean force_iterm2, gdouble gamma) {
     VideoPlayer *player = g_new0(VideoPlayer, 1);
     if (!player) {
         return NULL;
@@ -540,6 +541,7 @@ VideoPlayer* video_player_new(gint work_factor, gboolean force_text, gboolean fo
             .force_sixel = force_sixel,
             .force_kitty = force_kitty,
             .force_iterm2 = force_iterm2,
+            .gamma = gamma,
             .dither_mode = CHAFA_DITHER_MODE_NONE,
             .color_extractor = CHAFA_COLOR_EXTRACTOR_AVERAGE,
             .optimizations = CHAFA_OPTIMIZATION_REUSE_ATTRIBUTES